Remove library-tangle-file
This function was added so that I could use `C-c l t' to tangle the
file.  But I don't need a global key for that, since I only tangle the
file when it is already open.  And if that is the case, I can simply
use the key provided by org-babel itself: `C-c C-v f', or, `C-c C-v
C-f'.